completes basket functionality

+++ src/redux/actions/BasketItemsAction.js
@@ -56,7 +56,7 @@ export const addBasketItem = (id, payload, token) => (dispatch) => {
     });
 };
 
-export const deleteBasketItem = (basketItemId, token) => (dispatch) => {
+export const deleteBasketItem = (userId, basketId, basketItemId, token) => (dispatch) => {
   dispatch(basketItemLoading());
   headers = {
     "Content-Type": "application/json",
@@ -67,23 +67,25 @@ export const deleteBasketItem = (basketItemId, token) => (dispatch) => {
     .delete(`${baseUrl}/basket_items/${basketItemId}`, { headers })
     .then((response) => {
       dispatch(deleteBasketItemSuccess(response.data));
+      dispatch(getShoppingBasket(userId, basketId, token))
     })
     .catch((error) => {
       dispatch(deleteBasketItemError({ message: error.message }));
     });
 };
 
-export const deleteAllBasketItems = (basketId, token) => (dispatch) => {
-  dispatch(deleteAllBasketItemsLoading());
+export const deleteAllBasketItems = (userId, basketId, token) => (dispatch) => {
+  dispatch(basketItemLoading());
   headers = {
     "Content-Type": "application/json",
     Authorization: `Bearer ${token}`,
   };
 
   axios
-    .delete(`${baseUrl}/basket_items/${basketId}`, { headers })
+    .delete(`${baseUrl}/basket_items/basket/${basketId}`, { headers })
     .then((response) => {
       dispatch(deleteAllBasketItemsSuccess(response.data));
+      dispatch(getShoppingBasket(userId, basketId, token))
     })
     .catch((error) => {
       dispatch(deleteAllBasketItemsError({ message: error.message }));
